Remove isEnd() and isBeginning(); add hasCharacter()
Yesterday, I added the isEnd() and isBeginning() methods. However, I don't like
those names, and I think the methods themselves are of limited usefulness. I
really only care about whether or not the stream has characters at a point in
time - i.e., is it over or not. Now, you can do that with the hasCharacter()
method.