Use Jstewmc\Chunker library
Before, I tried to chunk and split in this library. It was getting to be too
much. So, I moved all the chunking code to the chunking library and made the
chunker a dependency. This is not backwards-compatible (since the constructor's
signature changed among other things). However, it's for the best. Now, this
class is much much simpler.